+++ src/app/cross/sass/globals.scss
@@ -20,7 +20,3 @@ body {
   overflow-y: hidden;
   overflow-x: hidden;
 }
-@media (max-width: 398px) {
-  body {
-  }
-}
