--- src/app/cross/ui/components/availableMoneyCard/index.scss
@@ -1,1 +1,0 @@
-@import './style/desktop.scss';
