+++ src/app/dashBoard/index.scss
@@ -1,1 +1,4 @@
+@import '../cross/sass/variables';
 @import './styles/desktop.scss';
+@import './styles/mobile.scss';
+@import './styles/tablet.scss';
