+++ src/app/dashBoard/perfil/page.js
@@ -1,0 +1,2 @@
+import Perfil from './ui/composite-wrappers/perfil/src/CWPerfil'
+export default Perfil
