+++ src/app/dashBoard/perfil/ui/composite-wrappers/perfil/src/CWPerfil.jsx
@@ -1,0 +1,13 @@
+'use client'
+import './styles/desktop.scss'
+
+const CWPerfil = () => {
+  const defaultClass = 'dashboard'
+  return (
+    <div>
+      <h1>Perfil</h1>
+    </div>
+  )
+}
+
+export default CWPerfil
