+++ src/app/dashBoard/perfil/ui/composite-wrappers/perfil/src/styles/desktop.scss
@@ -1,0 +1,1 @@
+@import '../../../../../../../cross/sass/variables';
