+++ src/app/dashBoard/styles/mobile.scss
@@ -1,0 +1,32 @@
+@media (max-width: 398px) {
+  .dashboard {
+    width: 100%;
+    min-height: 100vh;
+
+    &__containMenuDesplegable {
+      display: none;
+    }
+    &__content {
+      overflow-y: scroll;
+      width: 100%;
+      background-color: rgba(210, 255, 236, 1);
+    }
+    &__contain-buttons {
+      margin: 16px;
+      width: 360px;
+      height: 154px;
+      display: flex;
+      flex-direction: column;
+      justify-content: space-around;
+      align-items: center;
+    }
+
+    &__contain-input {
+      margin: 0px 16px 0px 16px;
+      &::placeholder {
+        background: rgba(0, 0, 0, 0.5);
+        size: 16px;
+      }
+    }
+  }
+}
