+++ src/app/dashBoard/styles/tablet.scss
@@ -1,5 +1,5 @@
-//Desktop Styles
-@media screen and (min-width: 835px) {
+//Tablet Styles
+@media screen and (min-width: 399px) and (max-width: 834px) {
   .dashboard {
     &__layout {
       display: flex;
@@ -8,31 +8,31 @@
       display: flex;
     }
     &__content {
-      max-height: calc(100vh - 120px);
+      max-height: calc(100vh - 20px);
       flex: 1;
-      padding: 16px 79px 0 79px;
+      padding: 68px 52px 0px 52px;
+      overflow-y: scroll;
       background-color: rgba(210, 255, 236, 1);
     }
 
-    &__contain-card-black {
-      height: 230px;
-    }
-
     &__contain-buttons {
-      margin: 16px;
+      width: 511px;
+      height: 190px;
       display: flex;
-      flex-direction: row;
+      flex-direction: column;
       justify-content: space-around;
-      gap: 20px;
-      height: 106px;
+      align-items: center;
     }
 
     &__contain-input {
       margin: 16px;
+      height: 64px;
 
       &::placeholder {
-        background: rgba(255, 255, 255, 1);
-        size: $heading-3;
+        // }
+      }
+
+      &__contain-activity {
       }
     }
   }
