+++ src/app/dashBoard/ui/components/availableMoneyCard/index.js
