+++ src/app/dashBoard/ui/components/availableMoneyCard/index.scss
@@ -1,0 +1,1 @@
+@import './styles/desktop.scss';
