+++ src/app/dashBoard/ui/components/availableMoneyCard/styles/desktop.scss
@@ -1,7 +1,7 @@
-@import '../../../../sass/variables';
+@import '../../../../../cross/sass/variables';
 
 @media (max-width: 398px) {
-  .container-availableMoneyCard {
+  .available-money-card {
     width: 100%;
     height: 160px;
     display: flex;
@@ -9,7 +9,7 @@
     justify-content: space-around;
     align-items: center;
 
-    &__views {
+    &__options {
       width: 170px;
       height: 16px;
       display: flex;
@@ -38,14 +38,14 @@
 }
 
 @media screen and (min-width: 399px) and (max-width: 834px) {
-  .container-availableMoneyCard {
+  .available-money-card {
     height: 230px;
     display: flex;
     flex-direction: column;
     justify-content: space-evenly;
     align-items: center;
 
-    &__views {
+    &__options {
       width: 170px;
       margin-left: 260px;
       display: flex;
@@ -71,14 +71,14 @@
 }
 
 @media screen and (min-width: 835px) {
-  .container-availableMoneyCard {
+  .available-money-card {
     width: 100%;
     height: 230px;
     display: flex;
     flex-direction: column;
     justify-content: space-evenly;
 
-    &__views {
+    &__options {
       width: 170px;
       height: 16px;
       display: flex;
