+++ src/app/dashBoard/ui/components/cardOptions/index.js
@@ -1,0 +1,3 @@
+import CardOptions from './src/CardOptions '
+
+export default CardOptions
