+++ src/app/dashBoard/ui/components/cardOptions/index.scss
@@ -1,0 +1,1 @@
+@import './styles/desktop.scss';
