--- src/app/dashBoard2/index.scss
@@ -1,1 +1,0 @@
-@import './ui/composite-wrappers/styles/desktop.scss';
