--- src/app/dashBoard2/page.js
@@ -1,2 +1,0 @@
-import DashBoard from './ui/composite-wrappers/CWDasboard/src/CWDasboard'
-export default DashBoard
