+++ src/app/dashBoard2/ui/composite-wrappers/CWDasboard/src/CWDasboard.jsx
@@ -3,9 +3,15 @@ import './styles/desktop.scss'
 
 import {useEffect, useState} from 'react'
 
+import Button from '../../../../../cross/ui/components/button'
+import InputSearch from '../../../../../cross/ui/components/inputSearch'
 import MenuDesplegable from '../../../../../cross/ui/components/menuDesplegable'
-import getDataLocalStore from '../../../../../cross/core/uses-cases/getDataLocalStore'
 import CardBlack from '../../../../../cross/ui/components/cardBlack'
+import CardActivity from '../../../../../cross/ui/components/cardActivity'
+import CardList from '../../../../../cross/ui/components/cardList'
+
+import getDataLocalStore from '../../../../../cross/core/uses-cases/getDataLocalStore'
+
 import AvailableMoneyCard from '../../../../../cross/ui/components/availableMoneyCard'
 
 const CWDasboard = () => {
@@ -13,26 +19,33 @@ const CWDasboard = () => {
   const [account, setAccount] = useState(0)
 
   useEffect(() => {
-    debugger
     setAccount(() => JSON.parse(getDataLocalStore('account')))
     console.log(account)
   }, [])
 
   return (
-    <main className={`${defaultClass}`}>
-      <div className={`${defaultClass}__layout`}>
-        <div className={`${defaultClass}__containMenuDesplegable`}>
-          <MenuDesplegable />
-        </div>
-        <div className={`${defaultClass}__content`}>
-          <div className={`${defaultClass}__contain-card-black`}>
-            <CardBlack size="AvailableAmountCard">
-              <AvailableMoneyCard />
-            </CardBlack>
-          </div>
-        </div>
+    <div className={`${defaultClass}__contain-card-black`}>
+      <CardBlack size="AvailableAmountCard">
+        <AvailableMoneyCard />
+      </CardBlack>
+
+      <div className={`${defaultClass}__contain-buttons`}>
+        <Button size="button-activity" label="Cargar dinero" color="green" />
+        <Button
+          size="button-activity"
+          label="Pago de servicios"
+          color="green"
+        />
       </div>
-    </main>
+      <div className={`${defaultClass}__contain-input`}>
+        <InputSearch size="medium-prompt" placeholder="Busca tu actividad" />
+      </div>
+      <div className={`${defaultClass}__contain-activity`}>
+        <CardActivity size="Activity">
+          <CardList />
+        </CardActivity>
+      </div>
+    </div>
   )
 }
