--- src/app/dashBoard2/ui/composite-wrappers/CWDasboard/src/styles/desktop.scss
@@ -1,153 +1,0 @@
-@import '../../../../../../cross/sass/variables';
-@media (max-width: 398px) {
-  .dashboard {
-    width: 100%;
-    min-height: 100vh;
-
-    &__layout {
-    }
-    &__containMenuDesplegable {
-      display: none;
-    }
-    &__content {
-      overflow-y: scroll;
-      width: 100%;
-
-      background-color: rgba(210, 255, 236, 1);
-    }
-
-    &__contain-header {
-      width: 78px;
-      height: 22px;
-      margin: 16px;
-      display: flex;
-      flex-direction: row;
-      justify-content: space-around;
-
-      p {
-        text-decoration: underline;
-      }
-    }
-
-    &__contain-cardBlack {
-    }
-
-    &__contain-buttons {
-      margin: 16px;
-      width: 360px;
-      height: 154px;
-      display: flex;
-      flex-direction: column;
-      justify-content: space-around;
-      align-items: center;
-    }
-
-    &__contain-input {
-      margin: 0px 16px 0px 16px;
-      &::placeholder {
-        background: rgba(0, 0, 0, 0.5);
-        size: 16px;
-      }
-    }
-
-    &__contain-activity {
-    }
-  }
-}
-
-//Tablet Styles
-@media screen and (min-width: 399px) and (max-width: 834px) {
-  .dashboard {
-    &__layout {
-      display: flex;
-    }
-    &__containMenuDesplegable {
-      display: flex;
-    }
-    &__content {
-      max-height: calc(100vh - 20px);
-      flex: 1;
-      padding: 68px 52px 0px 52px;
-      overflow-y: scroll;
-      background-color: rgba(210, 255, 236, 1);
-    }
-    &__contain-header {
-      display: none;
-
-      p {
-      }
-    }
-    &__contain-cardBlack {
-    }
-
-    &__contain-buttons {
-      width: 511px;
-      height: 190px;
-      display: flex;
-      flex-direction: column;
-      justify-content: space-around;
-      align-items: center;
-    }
-
-    &__contain-input {
-      margin: 16px;
-      height: 64px;
-
-      &::placeholder {
-        // }
-      }
-
-      &__contain-activity {
-      }
-    }
-  }
-}
-
-//Desktop Styles
-@media screen and (min-width: 835px) {
-  .dashboard {
-    &__layout {
-      display: flex;
-    }
-    &__containMenuDesplegable {
-      display: flex;
-    }
-    &__content {
-      max-height: calc(100vh - 120px);
-      flex: 1;
-      padding: 16px 79px 0 79px;
-      background-color: rgba(210, 255, 236, 1);
-    }
-
-    &__contain-header {
-      display: none;
-      p {
-      }
-    }
-
-    &__contain-cardBlack {
-      height: 230px;
-    }
-
-    &__contain-buttons {
-      margin: 16px;
-      display: flex;
-      flex-direction: row;
-      justify-content: space-around;
-      gap: 20px;
-      height: 106px;
-    }
-
-    &__contain-input {
-      margin: 16px;
-
-      &::placeholder {
-        background: rgba(255, 255, 255, 1);
-        size: $heading-3;
-      }
-    }
-
-    &__contain-activity {
-    }
-  }
-}
